requirements.txt
Django==1.10.4
django-compressor=2.1
drf-nested-routers==0.11.1
djangorestframework==3.5.3
dj-static==0.0.6
markdown==2.6.7
django-filter==1.0.1
pygments==2.1.3
